requirements.txt
langchain_core==0.3.5
langchain_openai==0.2.0
langchain_pinecone==0.2.0
langgraph==0.2.23
pinecone==5.3.1
pydantic==2.9.2
pymysql==1.1.1
python-dotenv==1.0.1
streamlit
google-api-python-client
protoc_gen_openapiv2
grpcio
cryptography
